backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-dotenv
pydantic

# NLP
scikit-learn
transformers
torch
sentencepiece
nltk

# Imbalance handling
imbalanced-learn

# OCR
pytesseract
Pillow

# Domain analysis
python-whois

# Utilities
httpx
python-multipart
pandas
numpy
joblib
faker
